+++ chart/spec/configuration/middleware_spec.rb
@@ -41,7 +41,7 @@ describe 'middleware configuration' do
           t.dig(
             'ConfigMap/test-registry',
             'data',
-            'config.yml'
+            'config.yml.tpl'
           ),
           permitted_classes: [Symbol]
         )['middleware']).to include(YAML.safe_load(%(